[linux-packaging-snippets] clang-conf-advanced: Implement advanced clang configuration
This feature deals with both apt and custom clang toolchains.
Also covers the toolchains for amd64 and arm64 hosts.

For amd64, the Droidian toolchains from apt will be used by default.
But easiest can be switched to Debian standard packages.

For arm64 hosts, the Debian apt toolchains will be used. Only clang 11, 13 and 14 are currently supported for this arch.

Finally, CUSTOM_CLANG feature is expanded. It automates a custom toolchain download, and sets the build path.

ADVICE: This implies moving the first clang custom code to the
kernel-snippet-clang-conf-advanced.mk

The toolchain needs to be on a git repo.
A sample repo with the clang 9, 10, 11, 12, 14 versions for amd64
can be found on:
https://github.com/android-berb/platform.prebuilts.clang.host.linux-x86

The prepared branches start with "clang-"

The clang custom download feature, required a shellscript code,
probably will be integrated in the releng tools.

To enable this feature, add to the debian/rules, after the kernel-snippet include:
include /usr/share/linux-packaging-snippets/kernel-snippet-clang-conf-advanced.mk

To show some extra documentation and config for kernek-info.mk,
a full sample from xiaomi vayu is included.

Signed-off-by: berbascum <berbascum@ticv.cat>